+++ spec/directions/bake_example_spec.rb
@@ -153,6 +153,97 @@ RSpec.describe Kitchen::Directions::BakeExample do
     end
   end
 
+  context 'when there is more than one exercise with solutions' do
+    let(:exercise) do
+      <<~HTML
+        <div data-type="exercise" id="exercise_id">
+          <div data-type="problem" id="problem_id">
+            <div data-type="title" id="title_id">Evaluating Functions</div>
+            <p>example content</p>
+          </div>
+          <div data-type="solution" id="solution_id">
+            <p>Solution One Content</p>
+          </div>
+          <div data-type="commentary" id="commentary_id">
+            <div data-type="title" id="title_id">Analysis</div>
+          </div>
+        </div>
+        <div data-type="exercise" id="exercise_id">
+          <div data-type="problem" id="problem_id">
+            <div data-type="title" id="title_id">Evaluating Functions</div>
+            <p>example content</p>
+          </div>
+          <div data-type="solution" id="solution_id">
+            <p>Solution Two Content</p>
+          </div>
+        </div>
+      HTML
+    end
+
+    it 'numbers the solutions correctly when numbered_solutions true' do
+      described_class.v1(
+        example: example,
+        number: 4,
+        title_tag: 'title-tag-name',
+        numbered_solutions: true
+      )
+      expect(example).to match_normalized_html(
+        <<~HTML
+          <div data-type="example" id="example-test">
+            <title-tag-name class="os-title">
+              <span class="os-title-label">Example </span>
+              <span class="os-number">4</span>
+              <span class="os-divider"> </span>
+            </title-tag-name>
+            <div class="body">
+              <h4 data-type="title">example title becomes h4</h4>
+              <p>content</p>
+              <div data-type="exercise" id="exercise_id" class="unnumbered">
+                <div data-type="problem" id="problem_id">
+                  <div class="os-problem-container">
+                    <h4 data-type="title" id="title_id">Evaluating Functions</h4>
+                    <p>example content</p>
+                  </div>
+                </div>
+                <div data-type="solution" id="solution_id">
+                  <h4 data-type="solution-title">
+                    <span class="os-title-label">Solution </span>
+                    <span class="os-number">1</span>
+                  </h4>
+                  <div class="os-solution-container">
+                    <p>Solution One Content</p>
+                  </div>
+                </div>
+                <div data-type="commentary" id="commentary_id">
+                  <h4 data-type="commentary-title" id="title_id">
+                    <span class="os-title-label">Analysis</span>
+                  </h4>
+                </div>
+              </div>
+              <div class="unnumbered" data-type="exercise" id="exercise_id">
+                <div data-type="problem" id="problem_id">
+                  <div class="os-problem-container">
+                    <h4 data-type="title" id="title_id">Evaluating Functions</h4>
+                    <p>example content</p>
+                  </div>
+                </div>
+                <div data-type="solution" id="solution_id">
+                  <h4 data-type="solution-title">
+                    <span class="os-title-label">Solution </span>
+                    <span class="os-number">2</span>
+                  </h4>
+                  <div class="os-solution-container">
+                    <p>Solution Two Content</p>
+                  </div>
+                </div>
+              </div>
+            </div>
+          </div>
+        HTML
+      )
+    end
+  end
+
   it 'stores info in the pantry' do
     expect { described_class.v1(example: example, number: 4, title_tag: 'title-tag-name') }.to change {
       example.pantry(name: :link_text).get(example.id)