+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 * Sort terms in `BakeChapterGlossary` in language specific way (major)
+* Spanish translation change (minor)
 
 ## [7.0.0] - 2021-06-21