+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [8.0.0] - 2021-06-29
+
 * Sort terms in `BakeChapterGlossary` in language specific way (major)
 * Spanish translation change (minor)
 * Fixed the implementation of `Element#wrap_children` to reuse existing document elements (major).